requirements.txt
pynacl~=1.5.0
yt-dlp~=2025.2.19
ffmpeg~=1.4
discord~=2.3.2
spotipy~=2.24.0
python-dotenv~=1.0.1